Security Camera Tracking Updates
Made the camera now follow the player with better logic. It now finds a better "Last known location" before losing sight of the player completely.

TODO: There's a bug where the detection of the player is blocked. This may actually be due to the collider of the camera interfering with the RangeSensor on it.
TODO: Fix the weird camera rotation issue. The stage exit camera does it sometimes. It rotates through the wall and the player can be spotted.